refactor!: move to zero-config return-based DeeplinkProcessor and generated module processor
