+++ deepmatch-testing/src/main/kotlin/com/aouledissa/deepmatch/processor/FakeDeeplinkProcessor.kt
@@ -3,7 +3,7 @@ package com.aouledissa.deepmatch.processor
 import android.net.Uri
 import com.aouledissa.deepmatch.api.DeeplinkParams
 
-class FakeDeeplinkProcessor : DeeplinkProcessor {
+class FakeDeeplinkProcessor : DeeplinkProcessor(specs = emptySet()) {
     private var resultProvider: (Uri) -> DeeplinkParams? = { null }
     private var lastDeeplink: Uri? = null
