+++ docs/gradle_plugin.md
@@ -16,6 +16,7 @@ The primary responsibilities and capabilities of the DeepMatch Gradle plugin are
     *   The plugin generates Kotlin source code, most notably a class named something like `[DeeplinkName]DeeplinkSpecs` (the exact name might vary based on the deeplink name in the `.yml` file).
     *   The plugin also generates params classes for deeplinks with typed path/query/fragment values.
     *   It additionally generates a module-level sealed interface (for example, module `app` -> `AppDeeplinkParams`) implemented by all generated params classes, enabling exhaustive `when` matching in app code.
+    *   It generates a module-level processor object (for example, module `app` -> `AppDeeplinkProcessor`) wired with all generated specs so no manual registration is required.
 
 4.  **Integration with Build Process:**
     *   The plugin hooks into the Android Gradle Plugin's build lifecycle.
@@ -60,6 +61,7 @@ During the build the plugin generates Kotlin sources under `build/generated/` an
 ### Generated Artifacts
 
 - `<ModuleName>DeeplinkParams.kt` — module-level sealed interface implemented by generated params classes.
+- `<ModuleName>DeeplinkProcessor.kt` — module-level processor object extending `DeeplinkProcessor` and preconfigured with all generated specs.
 - `*DeeplinkSpecs.kt` — exposes a `DeeplinkSpec` property per configuration entry.
 - `*DeeplinkParams.kt` — optional data class emitted when a deeplink defines typed template, query, or fragment parameters.
 - Generated manifest file — contains `<intent-filter>` definitions that Gradle merges into the final manifest.
